feat: funnel ReferenceError 해결
